infrastructure/requirements.txt
aws-cdk-lib==2.129.0
constructs>=10.0.0,<11.0.0
python-dotenv>=0.19.0